+++ lib/util/utils.ts
@@ -20,11 +20,14 @@ import {
   Data,
   getFilePosition,
   FilePosition,
-  getDescendantFilePosition
+  getDescendantFilePosition,
+  copyInfo
 } from "@ts-common/source-map"
-import { Suppression, SuppressionItem } from "@ts-common/azure-openapi-markdown"
+import { Suppression, SuppressionItem } from "@azure/openapi-markdown"
 import { splitPathAndReverse, isSubPath } from "./path"
 import jp = require("jsonpath")
+import { getSchemaObjectInfo, setSchemaInfo } from '../validators/specTransformer'
+import * as it from "@ts-common/iterator"
 
 export type DocCache = MutableStringMap<Promise<SwaggerObject>>
 
@@ -82,7 +85,12 @@ export async function parseJson(
     suppressionItems: ReadonlyArray<SuppressionItem>
   ): ReadonlyArray<SuppressionItem> => {
     const urlReversed = splitPathAndReverse(specPath)
-    return suppressionItems.filter(s => isSubPath(urlReversed, splitPathAndReverse(s.from)))
+    return suppressionItems.filter(
+      s => it.some(
+        it.isArray(s.from) ? s.from : [s.from],
+        from => isSubPath(urlReversed, splitPathAndReverse(from))
+      )
+    )
   }
 
   const suppressionArray =
@@ -97,6 +105,26 @@ export async function parseJson(
 
   const doc = docCache[specPath]
 
+  const applySuppression = (result: SwaggerObject) => {
+    const rootInfo = getFilePosition(result)
+    // apply suppression
+    for (const s of suppressionArray) {
+      if (s.where !== undefined) {
+        const paths = it.flatMap(
+          it.isArray(s.where) ? s.where : [s.where],
+          where => jp.paths(result, where)
+        )
+        for (const p of paths) {
+          // drop "$" and apply suppressions.
+          setSuppression(getDescendantFilePosition(result, drop(p)), s.suppress)
+        }
+      } else {
+        setSuppression(rootInfo, s.suppress)
+      }
+    }
+    return result
+  }
+
   if (doc) {
     return await doc
   }
@@ -110,7 +138,7 @@ export async function parseJson(
         "https://raw.githubusercontent.com$2$3"
       )
     }
-    const res = makeRequest({ url: specPath, errorOnNon200Response: true })
+    const res = makeRequest({ url: specPath, errorOnNon200Response: true }).then(applySuppression)
     docCache[specPath] = res
     return await res
   } else {
@@ -118,20 +146,7 @@ export async function parseJson(
     try {
       const fileContent = fs.readFileSync(specPath, "utf8")
       const result = parseContent(specPath, fileContent)
-      const rootInfo = getFilePosition(result)
-      // apply suppression
-      for (const s of suppressionArray) {
-        if (s.where !== undefined) {
-          const paths = jp.paths(result, s.where)
-          for (const p of paths) {
-            // drop "$" and apply suppressions.
-            setSuppression(getDescendantFilePosition(result, drop(p)), s.suppress)
-          }
-        } else {
-          setSuppression(rootInfo, s.suppress)
-        }
-      }
-      //
+      applySuppression(result)
       docCache[specPath] = Promise.resolve(result)
       return result
     } catch (err) {
@@ -746,6 +761,16 @@ export function allowNullType<T extends Entity>(
   entity: T,
   isPropRequired?: boolean | {}
 ): T {
+
+  const info = getSchemaObjectInfo(entity)
+
+  const nullable = () => {
+    const typeNull: SchemaObject = setSchemaInfo({ type: "null" }, info)
+    const typeArray = copyInfo(entity, [entity, typeNull])
+    const newEntity: SchemaObject = setSchemaInfo({ anyOf: typeArray }, info)
+    entity = newEntity as T
+  }
+
   // if entity has a type
   if (entity && entity.type) {
     // if type is an array
@@ -775,12 +800,13 @@ export function allowNullType<T extends Entity>(
       const savedEntity = entity
       // handling nullable parameters
       if (savedEntity.in) {
-        entity.anyOf = [{ type: entity.type }, { type: "null" }]
+        const typeNull: SchemaObject = setSchemaInfo({ type: "null" }, info)
+        const typeEntity: SchemaObject = setSchemaInfo({ type: entity.type }, info)
+        const typeArray: ReadonlyArray<SchemaObject> = copyInfo(entity, [typeEntity, typeNull])
+        entity.anyOf = typeArray
         delete entity.type
       } else {
-        entity = {
-          anyOf: [savedEntity, { type: "null" }]
-        } as any
+        nullable()
       }
     }
   }
@@ -791,10 +817,7 @@ export function allowNullType<T extends Entity>(
     entity.$ref &&
     shouldAcceptNullValue(entity["x-nullable"], isPropRequired)
   ) {
-    const savedEntity = entity
-    entity = {
-      anyOf: [savedEntity, { type: "null" }]
-    } as any
+    nullable()
   }
   return entity
 }
@@ -893,20 +916,18 @@ export function allowNullableParams(
  * @param {string} str - The string to be sanitized.
  * @returns {string} result - The sanitized string.
  */
-export function sanitizeFileName(str: string): string {
-  return str
+export const sanitizeFileName = (str: string): string =>
+  str
     ? str
         .replace(/[{}\[\]'";\(\)#@~`!%&\^\$\+=,\/\\?<>\|\*:]/gi, "")
         .replace(/(\s+)/gi, "_")
     : str
-}
 
 /**
  * Checks if the property is required in the model.
  */
-function isPropertyRequired(propName: unknown, model: SchemaObject) {
-  return model.required ? model.required.some(p => p === propName) : false
-}
+const isPropertyRequired = (propName: unknown, model: SchemaObject) =>
+  model.required ? model.required.some(p => p === propName) : false
 
 /**
  * Contains the reverse mapping of http.STATUS_CODES