+++ lib/validators/specResolver.ts
@@ -28,10 +28,9 @@ import {
 } from "@ts-common/string-map"
 import { resolveNestedDefinitions } from "./resolveNestedDefinitions"
 import { getOperations } from "../util/methods"
-import { transform } from "./specTransformer"
 import { map, toArray } from "@ts-common/iterator"
 import { arrayMap } from '@ts-common/source-map'
-import { Suppression } from '@ts-common/azure-openapi-markdown';
+import { Suppression } from '@azure/openapi-markdown';
 
 const ErrorCodes = C.ErrorCodes
 
@@ -210,9 +209,7 @@ export class SpecResolver {
       // resolve nested definitions
       this.specInJson = resolveNestedDefinitions(this.specInJson, this.options)
 
-      // all transformations without dependencies should be moved here)
-      this.specInJson = transform(this.specInJson)
-      // other resolvers
+      // other resolvers (should be moved to resolveNestedDefinitions())
       if (this.options.shouldResolveAllOf) {
         this.resolveAllOfInDefinitions()
       }