+++ lib/validators/suppressions.ts
@@ -2,18 +2,18 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 import * as path from "path"
-import * as fs from "fs"
 import * as md from "@ts-common/commonmark-to-markdown"
-import * as amd from "@ts-common/azure-openapi-markdown"
-import { isArray } from '@ts-common/iterator';
+import * as amd from "@azure/openapi-markdown"
+import { isArray } from "@ts-common/iterator"
+import * as vfs from "@ts-common/virtual-fs"
 
-export const getSuppressions = (specPath: string): undefined | amd.Suppression => {
+export const getSuppressions = async (specPath: string): Promise<undefined | amd.Suppression> => {
   // find readme.md
-  const readMe = findReadMe(path.dirname(specPath))
+  const readMe = await amd.findReadMe(path.dirname(specPath))
   if (readMe === undefined) {
     return undefined
   }
-  const readMeStr = fs.readFileSync(readMe).toString()
+  const readMeStr = await vfs.readFile(readMe)
   const cmd = md.parse(readMeStr)
   const suppressionCodeBlock = amd.getCodeBlocksAndHeadings(cmd.markDown).Suppression
   if (suppressionCodeBlock === undefined) {
@@ -24,19 +24,4 @@ export const getSuppressions = (specPath: string): undefined | amd.Suppression =
     return undefined
   }
   return suppression
-}
-
-const findReadMe = (dir: string): string | undefined => {
-  dir = path.resolve(dir)
-  while (true) {
-    const fileName = path.join(dir, "readme.md")
-    if (fs.existsSync(fileName)) {
-      return fileName
-    }
-    const newDir = path.dirname(dir)
-    if (newDir === dir) {
-      return undefined
-    }
-    dir = newDir
-  }
 }