+++ lib/wireFormatGenerator.ts
@@ -21,7 +21,8 @@ import { PathTemplateBasedRequestPrepareOptions } from "ms-rest"
 import { Responses, Headers } from "./templates/httpTemplate"
 import { map, toArray } from "@ts-common/iterator"
 import { getSuppressions } from './validators/suppressions';
-import { Suppression } from '@ts-common/azure-openapi-markdown';
+import { Suppression } from '@azure/openapi-markdown';
+import { setMutableProperty } from '@ts-common/property-set';
 
 const ErrorCodes = C.ErrorCodes
 
@@ -79,7 +80,7 @@ export class WireFormatGenerator {
       utils.clearCache()
     }
     try {
-      const suppression = getSuppressions(this.specPath)
+      const suppression = await getSuppressions(this.specPath)
       const result = await utils.parseJson(suppression, this.specPath)
       this.specInJson = result
       const specOptions = {
@@ -175,14 +176,15 @@ export class WireFormatGenerator {
   private constructErrorObject(
     code: unknown, message: string, innerErrors: Array<unknown>, _?: boolean
   ) {
-    const err = {
+    const err: {
+      code: unknown
+      message: string
+      innerErrors?: Array<unknown>
+    } = {
       code,
-      message,
-      innerErrors: undefined as (Array<unknown> | undefined)
+      message
     }
-    if (innerErrors) {
-      err.innerErrors = innerErrors
-    }
+    setMutableProperty(err, "innerErrors", innerErrors ? innerErrors : undefined)
     // if (!skipValidityStatusUpdate) {
     // this.updateValidityStatus();
     // }