+++ test/modelValidatorTests.ts
@@ -5,7 +5,6 @@
 
 import assert from "assert"
 import * as validate from "../lib/validate"
-import { SerializedError } from "../lib/util/baseValidationError"
 import { ModelValidator } from '../lib/validators/modelValidator';
 
 const specPath =
@@ -24,7 +23,7 @@ describe("Model Validation", () => {
         `swagger "${specPath}" with operation "${operationIds}" ` +
           `contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass with pretty format", async () => {
@@ -38,7 +37,7 @@ describe("Model Validation", () => {
         `swagger "${specPath}" with operation "${operationIds}" ` +
           `contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should fail with collapsed similar array elements errors", async () => {
@@ -57,12 +56,18 @@ describe("Model Validation", () => {
         `swagger "${specPath} with operation "${operationIds}" should report only 1 error.`
       )
 
+      if (result[0].errorDetails === undefined) {
+        throw new Error("result[0].errorDetails === undefined")
+      }
+      if (result[0].errorDetails.similarPaths === undefined) {
+        throw new Error("result[0].errorDetails.similarPaths === undefined")
+      }
       assert(
-        (result[0].errorDetails as SerializedError).similarPaths
+        result[0].errorDetails.similarPaths
           .length === 1,
         `swagger "${specPath} with operation "${operationIds}" error should have a similar path.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for paths in x-ms-paths with question mark", async () => {
@@ -74,7 +79,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for paths with quotes", async () => {
@@ -86,7 +91,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should fail for paths with path parameter value resulting in duplicate forward slashes", async () => {
@@ -99,7 +104,7 @@ describe("Model Validation", () => {
           result.length !== 0,
           `swagger "${specPath}" with operation "${operationIds}" contains passed incorrectly.`
         )
-        console.log(result)
+        // console.log(result)
       } catch (err) {
         assert.strictEqual(err.code, "REQUEST_VALIDATION_ERROR")
         assert.strictEqual(err.innerErrors[0].code, "DOUBLE_FORWARD_SLASHES_IN_URL")
@@ -118,7 +123,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for Activity_Dictionary", async () => {
@@ -131,7 +136,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for CircularAnimal_List", async () => {
@@ -144,7 +149,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should fail for CircularAnimal_IncorrectSibling_List", async () => {
@@ -214,7 +219,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -228,7 +233,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -238,12 +243,12 @@ describe("Model Validation", () => {
       const result = await validate.validateExamples(specPath2, undefined, {
         consoleLogLevel: "off"
       })
-      console.dir(result, { depth: null })
+      // console.dir(result, { depth: null })
       assert(
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should validate the presence of parameters", async () => {
@@ -251,12 +256,12 @@ describe("Model Validation", () => {
       const result = await validate.validateExamples(specPath2, undefined, {
         consoleLogLevel: "off"
       })
-      console.dir(result, { depth: null })
+      // console.dir(result, { depth: null })
       assert(
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -271,7 +276,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for formatInDefinition_Get", async () => {
@@ -284,7 +289,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for enumInResponse_Get", async () => {
@@ -297,7 +302,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for readOnlyProp_Get", async () => {
@@ -310,7 +315,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for arrayInResponse_List", async () => {
@@ -323,7 +328,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for objectInResponse_Get", async () => {
@@ -336,7 +341,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for typeArrayInResponse_Get", async () => {
@@ -349,7 +354,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for xnullableFalse_Get", async () => {
@@ -362,7 +367,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for requiredProp_Get", async () => {
@@ -375,7 +380,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for inlineResponse_Get", async () => {
@@ -388,7 +393,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for RefWithNullableAtTopLevelOperation_Get", async () => {
@@ -401,7 +406,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for definitionWithReference_Get", async () => {
@@ -414,7 +419,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for definitionWithReferenceNull_Get", async () => {
@@ -427,7 +432,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for definitionWithReferenceNotNullableOperation_Get", async () => {
@@ -440,7 +445,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
 
     it("should pass for nullableTopLevel_Get", async () => {
@@ -453,7 +458,7 @@ describe("Model Validation", () => {
         result.length === 0,
         `swagger "${specPath2}" with operation "${operationIds}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -463,12 +468,12 @@ describe("Model Validation", () => {
       const result = await validate.validateExamples(specPath2, undefined, {
         consoleLogLevel: "off"
       })
-      console.dir(result, { depth: null })
+      // console.dir(result, { depth: null })
       assert(
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -478,12 +483,12 @@ describe("Model Validation", () => {
       const result = await validate.validateExamples(specPath2, undefined, {
         consoleLogLevel: "off"
       })
-      console.dir(result, { depth: null })
+      // console.dir(result, { depth: null })
       assert(
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 
@@ -493,12 +498,12 @@ describe("Model Validation", () => {
       const result = await validate.validateExamples(specPath2, undefined, {
         consoleLogLevel: "off"
       })
-      console.dir(result, { depth: null })
+      // console.dir(result, { depth: null })
       assert(
         result.length === 0,
         `swagger "${specPath2}" contains model validation errors.`
       )
-      console.log(result)
+      // console.log(result)
     })
   })
 })